move egress_socket_select into net-utils/src/multihomed_sockets.rs
